+++ tests/mocks/ExampleFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace tests\mocks;
+
+final class ExampleFactory
+{
+    public function __invoke($class, ...$arguments)
+    {
+        return new $class(...$arguments);
+    }
+}